tutorials/multi-agent/requirements.txt
python-dotenv
smolagents
smolagents[gradio]
semantic-kernel
arize-phoenix 
opentelemetry-sdk 
opentelemetry-exporter-otlp 
openinference-instrumentation-smolagents